+++ package.json
@@ -6,7 +6,6 @@
     "localforage": "^1.5.5",
     "lodash.padend": "^4.6.1",
     "normalize.css": "^7.0.0",
-    "prop-types": "^15.6.0",
     "ramda": "^0.25.0",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
